speed up initial connect for new clients
When initially connecting to the network, the claiming cycle will have 15
iterations with a timeout of 300ms each. Should clients roam during this period
of time, the connectivity is lost and a multi-minute recovery-cycle will have
to commence. With this patch, clients do not have to wait out the claiming
cycle including the respective timeouts and retries any more. Netlink is
queried for the host-route of the special-IP of the client. If it does not
exist, then it is assumed that the client didn't roam but newly joined. The
whole claiming cycle is skipped and the special node-IP is directly assigned,
speeding up initial claiming for newly joined clients by 5 seconds.